refactor: 'NoteContentHeader' component, move all buttons to separate files
